Split firm agent modules
